Began Search Box refactor
